added treats appearing on every tap
implemented via object pooling. some treats are still missed when a user
taps but no bones are active but it is near perfect. there is an error
whenever a new treat object is created. this error occurs because no
objects in Unity/the inspector of the new treat objects are attached to
a game object.