Changed canvas layout, sprite aura script
The canvas is set to "Constant Pixel Size" so it won't resize on
different resolutions. Need to change the canvas so that it will work
with "Scale with screen size".
Changing the way sprite aura works so that it can work with sprite speed
ups on every third upgrade and at the max upgrade.