Fixed Double Sprite Bug On Tap
- Double sprite was happening after a few taps because DogTreats.cs and
TreatTapSprite2.cs were both accounting for a tap on the screen. Also,
both the parent and it's first child were being activated on tap. Fixed
by adding and if/else so that the parent and child cannot both be
activated at the same time; if the parent is active then it will only
call on TreatPool() instead of activating the parent again.
- This is a very rough summary. Good luck future Thien.